Attempt at secondary selection through providing sequential ID as vertex attribute.
This seems to work OK for points and lines.

For triangles flat shading is used so what comes through is the id of
provoking vertex (last in ES-3, this is configurable in GL-4 to be either the
first or the last vertex).

The implementation is rather messy as we need a new path in MeshRenderer.

gl_PrimitiveID exists in GL-4 and ES-3.2

gl_VertexID exists in ES-3.0 - this could effectively mimic what I have here.

gl_InstanceID is another option ... but then one would have to change the
shaders completely.